pyproject.toml
[project]
name = "pdf2xls"
version = "0.1.0"
description = "Transformar PDFs em planilhas"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "camelot-py>=1.0.0",
    "duckdb-engine>=0.15.0",
    "ghostscript>=0.7",
    "matplotlib>=3.10.1",
    "opencv-python>=4.11.0.86",
    "pymssql>=2.3.2",
    "sqlalchemy>=2.0.40",
    "unidecode>=1.3.8",
]
